services_openaitesting_1740662240.533000/requirements.txt
functions-framework
google-cloud-aiplatform
google-auth
requests
openai